Handle serialization of bool and int values
It is sometimes useful to be able to save props which are in various other
native Python types. Even though we can't and/or don't attempt to detect and
parse bools, ints and floats back into their native types it is still
convenient to be able to save properties which aren't plain strings.